requirements.txt
Flask>=2.0
tinydb>=4.7
python-dateutil>=2.8
watchdog